+++ src/truncate.ts
@@ -2,12 +2,15 @@
 // @ts-ignore
 import pg from "pg"
 
-export default async function truncate(DreamApplication: any, connectionName: string = 'default') {
+export default async function truncate(
+  DreamApp: any,
+  connectionName: string = "default"
+) {
   // this was only ever written to clear the db between tests,
   // so there is no way to truncate in dev/prod
   if (process.env.NODE_ENV !== "test") return false
 
-  const dreamconf = DreamApplication.getOrFail()
+  const dreamconf = DreamApp.getOrFail()
 
   // prior to @rvoh/dream@1.5.0, dbCredentials would point to the app's
   // root db credentials. After 1.5.0, this has been switched to be a record
@@ -15,9 +18,14 @@ export default async function truncate(DreamApplication: any, connectionName: st
   // To maintain backwards compatibility with older versions of dream, we will check
   // for the 'default' key, and if it exists, we will use the connectionName to
   // drill in, and otherwise fall back, since we must be in an older version of dream.
-  const credentials = dreamconf.dbCredentials[connectionName]?.primary || dreamconf.dbCredentials.primary
+  const credentials =
+    dreamconf.dbCredentials[connectionName]?.primary ||
+    dreamconf.dbCredentials.primary
 
-  if (!credentials) throw new Error(`Failed to locate db credentials for connectionName: ${connectionName}`)
+  if (!credentials)
+    throw new Error(
+      `Failed to locate db credentials for connectionName: ${connectionName}`
+    )
 
   const client = new pg.Client({
     host: credentials.host || "localhost",
@@ -43,7 +51,7 @@ LOOP
 END LOOP;
 END;
 $$;
-`,
+`
   )
   await client.end()
 }