+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## 1.2.0
+
+- update `provideDreamViteMatchers` to take `DreamApp` so `toMatchDreamModel(s)` doesn't use the deprecated `isDreamInstance`
+
 ## 1.1.1
 
-- add support for multiple connections when truncating. provide connectionName as a second argument when truncating. defaults to "default". Will maintain backwards compatibility with older versions of dream. 
+- add support for multiple connections when truncating. provide connectionName as a second argument when truncating. defaults to "default". Will maintain backwards compatibility with older versions of dream.
 
 ## 1.1.0